fix(auth): remove legacy dashboard session helper
